Move more tests back to their original file
