+++ multisensor_pipeline/tests/test_signal.py
@@ -2,6 +2,7 @@ from time import sleep
 
 from multisensor_pipeline.modules import QueueSink
 from multisensor_pipeline.modules.npy import RandomArraySource
+from multisensor_pipeline.modules.signal.filtering import OneEuroProcessor
 from multisensor_pipeline.modules.signal.sampling import DownsamplingProcessor
 from multisensor_pipeline.pipeline.graph import GraphPipeline
 
@@ -43,3 +44,71 @@ def test_down_sampling_processor_no_downsampling():
     assert 10 <= sink_1.queue.qsize()
     # No sink should not have received much more messages than the other
     assert abs(sink_0.queue.qsize() - sink_1.queue.qsize()) <= 4
+
+
+def test_down_sampling_processor_strong():
+    # Mock a setup like so:
+    # source -> processor -> sink
+
+    # (1) define the modules
+    source = RandomArraySource(
+        shape=None,
+        sampling_rate=100,
+        max_count=100,
+    )
+    processor = DownsamplingProcessor(sampling_rate=1)
+    sink = QueueSink()
+
+    # (2) add module to a pipeline...
+    pipeline = GraphPipeline()
+    pipeline.add_source(source)
+    pipeline.add_processor(processor)
+    pipeline.add_sink(sink)
+    # (3) ...and connect the modules
+    pipeline.connect(source, processor)
+    pipeline.connect(processor, sink)
+
+    # Test
+    pipeline.start()
+    sleep(1)
+    pipeline.stop()
+
+    # Assert
+    assert sink.queue.qsize() <= 3
+
+
+def test_one_euro_filter():
+    # Mock a setup like so:
+    # sink_0 <- source -> processor -> sink_1
+
+    # (1) define the modules
+    source = RandomArraySource(
+        shape=(2,),
+        sampling_rate=100,
+        max_count=10,
+    )
+    processor = OneEuroProcessor(
+        signal_topic_name="random",
+        signal_key="value",
+    )
+    sink_1 = QueueSink()
+    sink_0 = QueueSink()
+
+    # (2) add module to a pipeline...
+    pipeline = GraphPipeline()
+    pipeline.add_source(source)
+    pipeline.add_processor(processor)
+    pipeline.add_sink(sink_1)
+    pipeline.add_sink(sink_0)
+    # (3) ...and connect the modules
+    pipeline.connect(source, sink_0)
+    pipeline.connect(source, processor)
+    pipeline.connect(processor, sink_1)
+
+    # Test
+    pipeline.start()
+    sleep(2)
+    pipeline.stop()
+
+    # Assert
+    assert abs(sink_1.queue.qsize() - sink_0.queue.qsize()) <= 2
